change_cursor solve index overflow

+++ frame/select_filename.py
@@ -56,13 +56,15 @@ class SelectFilenameFrame(object):
             messagebox.showerror("Error", f"An error occurred: {str(e)}")
 
     def change_cursor(self, curent_cursor):
-        if ((self.last_item != -1) and (self.listbox_files_dataset.get(self.last_item)[0] == '>')):
-            selected_file = self.listbox_files_dataset.get(self.last_item)
+        if ((self.last_item != -1) and 
+            (self.last_item < self.listbox_files_dataset.size()) and 
+            (str(self.listbox_files_dataset.get(self.last_item)[0]) == '>')):
+            selected_file = str(self.listbox_files_dataset.get(self.last_item))
             self.listbox_files_dataset.insert(self.last_item, selected_file[len('> '):])
             self.listbox_files_dataset.delete(self.last_item+1)
             
         if (curent_cursor != -1):
-            selected_file = self.listbox_files_dataset.get(curent_cursor)
+            selected_file = str(self.listbox_files_dataset.get(curent_cursor))
             self.listbox_files_dataset.insert(curent_cursor, '> ' + selected_file)
             self.listbox_files_dataset.delete(curent_cursor+1)
         self.last_item = curent_cursor
